+++ src/ConnectionPoolFactory/SentinelConnectionPoolFactory.php
@@ -15,18 +15,21 @@ class SentinelConnectionPoolFactory implements ConnectionPoolFactory
 
     private float $timeout;
 
+    private ?float $operationTimeout;
+
     private ?int $retryWait;
 
     private ?int $maxFails;
 
     private bool $writeToReplicas;
 
-    public function __construct(array $sentinels, string $clusterId, int $database = 0, float $timeout = 0.0, ?int $retryWait = null, ?int $maxFails = null, bool $writeToReplicas = true)
+    public function __construct(array $sentinels, string $clusterId, int $database = 0, float $timeout = 0.0, ?int $retryWait = null, ?int $maxFails = null, bool $writeToReplicas = true, ?float $operationTimeout = null)
     {
         $this->sentinels = $sentinels;
         $this->clusterId = $clusterId;
         $this->database = $database;
         $this->timeout = $timeout;
+        $this->operationTimeout = $operationTimeout;
         $this->retryWait = $retryWait;
         $this->maxFails = $maxFails;
         $this->writeToReplicas = $writeToReplicas;
@@ -34,7 +37,7 @@ class SentinelConnectionPoolFactory implements ConnectionPoolFactory
 
     public function create(Driver $driver): SentinelConnectionPool
     {
-        $connectionPool = new SentinelConnectionPool($driver, $this->sentinels, $this->clusterId, $this->database, $this->timeout);
+        $connectionPool = new SentinelConnectionPool($driver, $this->sentinels, $this->clusterId, $this->database, $this->timeout, $this->operationTimeout);
         $connectionPool->setWriteToReplicas($this->writeToReplicas);
         if ($this->retryWait) {
             $connectionPool->setRetryWait($this->retryWait);