+++ src/ConnectionPoolFactory/SingleNodeConnectionPoolFactory.php
@@ -15,18 +15,21 @@ class SingleNodeConnectionPoolFactory implements ConnectionPoolFactory
 
     private float $timeout;
 
+    private ?float $operationTimeout;
+
     private bool $autoSelectDb;
 
     private ?int $retryWait;
 
     private ?int $maxFails;
 
-    public function __construct(string $host, int $port, int $database = 0, float $timeout = 0.0, bool $autoSelectDb = true, ?int $retryWait = null, ?int $maxFails = null)
+    public function __construct(string $host, int $port, int $database = 0, float $timeout = 0.0, bool $autoSelectDb = true, ?int $retryWait = null, ?int $maxFails = null, ?float $operationTimeout = null)
     {
         $this->host = $host;
         $this->port = $port;
         $this->database = $database;
         $this->timeout = $timeout;
+        $this->operationTimeout = $operationTimeout;
         $this->autoSelectDb = $autoSelectDb;
         $this->retryWait = $retryWait;
         $this->maxFails = $maxFails;
@@ -34,6 +37,6 @@ class SingleNodeConnectionPoolFactory implements ConnectionPoolFactory
 
     public function create(Driver $driver): SingleNodeConnectionPool
     {
-        return new SingleNodeConnectionPool($driver, $this->host, $this->port, $this->database, $this->timeout, $this->autoSelectDb, $this->retryWait, $this->maxFails);
+        return new SingleNodeConnectionPool($driver, $this->host, $this->port, $this->database, $this->timeout, $this->autoSelectDb, $this->retryWait, $this->maxFails, $this->operationTimeout);
     }
 }